app/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
aniso8601==9.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
bs4==0.0.1
cachelib==0.13.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
ecdsa==0.19.0
eventlet==0.36.1
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
http_ece==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
msgspec==0.18.6
multidict==6.1.0
packaging==24.1
propcache==0.2.0
psycopg2-binary==2.9.6
py-vapid==1.9.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
pywebpush==2.0.1
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
typing_extensions==4.12.2
Unidecode==1.3.7
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.15.2
